Add files via upload
I messed around with various different tools that was featured in the blot docs. I saw that there was the ability to upload svgs, so I drew up an image and uploaded it. As of right now, my goal is to simplify the drawing (which I have started) and make it so that certain features of the frog change upon generation.  I have also attached the frog I drew.